Extracted Rule Configuration structs to own files.
