+++ Source/SwiftLintFramework/Models/ConfigurationError.swift
@@ -1,0 +1,13 @@
+//
+//  ConfigurationError.swift
+//  SwiftLint
+//
+//  Created by Scott Hoyt on 1/19/16.
+//  Copyright © 2016 Realm. All rights reserved.
+//
+
+import Foundation
+
+public enum ConfigurationError: ErrorType {
+    case UnknownConfiguration
+}
